refactor(ui): extract thread scroll/status hooks into chrome/thread/scrolling.ts
Third extraction (Task 1): useMessageWindow (ENG-218 windowing) and
useStickToBottom (ENG-213 stick/jump) move out of vendo-thread.tsx.
Behavior-preserving.
